+++ src/pages/index.ts
@@ -2,6 +2,7 @@
  * Pages barrel export
  */
 
+export { default as HomePage } from './HomePage';
 export { default as DatasetOverviewPage } from './DatasetOverviewPage';
 export { default as NodeDetailPage } from './NodeDetailPage';
 export { default as EdgeDetailPage } from './EdgeDetailPage';